requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
DateTime==4.3
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.1
django-allauth==0.47.0
django-bootstrap3==21.2
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.7.1
et-xmlfile==1.1.0
filelock==3.4.2
idna==3.3
MarkupPy==1.14
numpy==1.22.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.5
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0